chimera: Detect races in directory deletion
Motivation:

A concurrent entry creation while deleting a directory may cause a non-empty
directory to be deleted. The result is a file with an entry in t_dirs but with
no parent in t_inodes.

This particulary seems to happen when FTP doors fail with 'no write pool
configured'; I suspect the doors or clients retry the creation while the SRM at
some point delete the upload directory.

Modification:

Check if an inodes was actually deleted and assume that if not it is because
the link count was not zero (which for directories means the directory was
not empty).

Result:

Fewer chances for inconsistencies in Chimera.

Target: 2.13
Request: 2.12
Request: 2.11
Request: 2.10
Require-notes: yes
Require-book: no
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: https://rb.dcache.org/r/8817/