chimera: Fix Chimera inconsistencies
Motivation:

Our (NDGF) Chimera instance shows a number of inconsistencies:

- t_dirs entries with non-existing parent

- t_inodes with no entries in t_dirs

- wrong link counts on directories

These are likely a result of the race conditions fixed in 2.14.

Modification:

Add changesets to move the lost inodes or lost directory entries to a newly
created lost+found directory in the chimera root. This directory is created
with a fixed PNFS ID just like the root directory.

Add changeset to add a foreign key constraint from the t_dirs parent to
t_inodes.  This prevents some of the inconsistencies from reappearing.

Add changeset to correct the link count on directories. This needs some
experimentation on a real world Chimera db to see if it can be optimized.

Unit tests have been adjusted to account for the pressence of a lost+found
directory.

Result:

Upon upgrade a lost+found directory is created, lost files will be moved to
lost+found and a foreign key constraint is added. On large instances these
changes may take a while to apply.

Target: trunk
Require-notes: yes
Require-book: no
Acked-by: Paul Millar <paul.millar@desy.de>
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: https://rb.dcache.org/r/8821/