security: drop AuthType from Origin
The Origin principal contains two pieces of information: the IP address
of the client and whether the authentication was "weak" or "strong".
The latter information is poorly defined and only used to distinguish
between authenticated users and anonymous users.

This patch selects between whether `AUTHENTICATED` or `ANONYMOUS` ACEs
apply by using the `Subjects#isNobody` method; i.e., whether or not the
user has a `UidPrincipal`.

Under certain circumstances, this patch has some user observable effects
that may require admin intervention, depending on dCache configuration.

Due to a bug in earlier versions, any Subject that has no `Origin`
principal is subject to both `AUTHENTICATED` and `ANONYMOUS` ACEs.  This
is clearly wrong as RFC 3530 states these classes of user are mutually
exclusive.

As specific examples of this, the hardcoded identities `Subjects.ROOT`
and `Subjects.NOBODY` do not contain an `Origin` principal, so may
suffer from this problem.  Currently, explicit `Subjects#isRoot` tests
bypass the ACL-checking so this is not a problem for `Subjects.ROOT`,
but is a problem for unmodified `Subjects.NOBODY` identities.  This
entry is used when:

    authenticating a user with the kauth file with an entry that
        starts "passwd " and has a password of '-',

    the xrootd door is configured to map a user to nobody (see below),

    the nfs door has a problem contacting gPlazma,

    the nfs mover is contacted with GSS authentication,

    the srm door is processing stored activity of a user that does
        not exist.

Additionally:

    the nfs door does not add the Origin principal to any users, so all
        non-root users are subject to both `AUTHENTICATED` and
        `ANONYMOUS` ACEs.

    the webdav, ftp and xrootd doors adds an Origin principal that
        states the user is subject to `AUTHENTICATED` ACEs even if the
        user accesses dCache without authentication; e.g.,
        `webdav.authz.anonymous-operations` is `FULL` or `READONLY`,
        `xrootd.authz.user` is `nobody`, ftp where user is mapped via
        kauth entry with a password entry of '-', and so on.

With this patch, the application of `AUTHENTICATED` and `ANONYMOUS` ACEs
is based on a common policy, independent of the door: the former is used
whenever the mapped user has a UidPrincipal; the latter whenever the
user has no UidPrincipal.

There are situations where the desirability of this policy is
questionable; for example, users being subject to `AUTHENTICATED` ACEs
for trusted-host authentication (client supplies the users identity;
e.g., NFS) or when all users are mapped to a specific identity (NFS and
xrootd).  This patch makes the tacid assumption that, under these
circumstances, the admin has ensured other mechanisms (e.g. firewalls,
careful client-side configuration) so the identity of the user is well
established and `AUTHENTICATED` is the correct ACE to use.

Target: master
Requires-notes: true
Requires-book: true
Patch: https://rb.dcache.org/r/8824/
Acked-by: Gerd Behrmann
Acked-by: Tigran Mkrtchyan
Acked-by: Dmitry Litvintsev