chimera: Fix deadlock
Motivation:

Chimera suffers from a deadlock between ChimeraNameSpaceProvider#setFileAttributes and
JdbcFs#remove:

2015-12-01 12:05:31 CET [25700] [[unknown] 131.169.192.35(40620)]: [1-1] ERROR:  deadlock detected
2015-12-01 12:05:31 CET [25700] [[unknown] 131.169.192.35(40620)]: [2-1] DETAIL:  Process 25700 waits for ShareLock on transaction 302154760; blocked by process 25404.
	Process 25404 waits for ShareLock on transaction 302154759; blocked by process 25700.
	Process 25700: DELETE FROM t_inodes WHERE ipnfsid=$1 AND inlink = 0
	Process 25404: UPDATE t_inodes SET ictime=$1,igeneration=igeneration+1,isize=$2,imtime=$3,iaccess_latency=$4,iretention_policy=$5 WHERE ipnfsid=$6 AND itype = 32768
2015-12-01 12:05:31 CET [25700] [[unknown] 131.169.192.35(40620)]: [3-1] HINT:  See server log for query details.
2015-12-01 12:05:31 CET [25700] [[unknown] 131.169.192.35(40620)]: [4-1] STATEMENT:  DELETE FROM t_inodes WHERE ipnfsid=$1 AND inlink = 0
2015-12-01 12:05:31 CET [25404] [[unknown] 131.169.192.35(40600)]: [1-1] LOG:  duration: 999.727 ms  execute S_12: UPDATE t_inodes SET ictime=$1,igeneration=igeneration+1,isize=$2,imtime=$3,iaccess_latency=$4,iretention_policy=$5 WHERE ipnfsid=$6 AND itype = 32768
2015-12-01 12:05:31 CET [25404] [[unknown] 131.169.192.35(40600)]: [2-1] DETAIL:  parameters: $1 = '2015-12-01 12:05:30.252+01', $2 = '213', $3 = '2015-12-01 12:05:30.252+01', $4 = '0', $5 = '0', $6 = '00001E50D35C2CBF480883E9A344191760D3'

These are from an attempt to concurrent set file attributes after a completed upload
(from the mover) and the client trying to delete that file (presumably the upload
failed so the client does a delete concurrently with the pool setting the file
attributes).

The deadlock is as follows. Call the thread calling setFileAttributes thread A, and
call the thread calling remove thread B (it is not a requirement that they are in
the same process).

Thread A iterates over the attribues of the FileAttribute object and collects those
that are stored in t_inodes in a Stat object. Locations, checksums and flags are
written to the DB right away. In particular the update from the pool contains
a location and likely also a checksum. These are inserted into t_locationinfo and
t_checksums; due to referential integrity constraints the corresponding row in
t_inodes receives a FOR KEY SHARE lock.

Thread B now begins deleting the same file; it removes the directory entry from
t_dirs following by decrementing the link count in t_inodes - this succeeds because
it doesn't update the key of the row. Thread B acquires a FOR NO KEY UPDATE lock as
a result. Thread B now executes the 'delete if link count is zero' statement. This
requires a full FOR UPDATE lock on the row, but because thread A got a FOR KEY SHARE
LOCK, this delete blocks.

Thread A now continues and tries to apply the attributes updates for t_inodes. This
requires a FOR NO KEY UPDATE lock, but it cannot get that because thread B already
holds that lock. Boom - deadlock.

This deadlock only occurs with PostgreSQL 9.3 and newer. Earlier versions only had
two types of row level locks and the scenario described doesn't work in that case.

Modification:

The patch updates the setFileAttributes method to update t_inodes before inserting
locations, flags and checksum in other tables. The update will acquire a
FOR NO KEY UPDATE lock on the t_inodes row and thus no FOR KEY SHARE lock is
acquired first. This elliminates the deadlock described above.

Result:

Hopefully one less deadlock, but a future patch should probably add generic retry
support in case of conflicting transactions (that's best practice for any RDBMS
accessing code).

Target: trunk
Request: 2.14
Require-notes: yes
Require-book: no
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: https://rb.dcache.org/r/8835/
(cherry picked from commit ade698a21f36fb070d7830b7e7712d6f724c691d)