doors: send Origin when logging in user
The Origin principal describes where from the client comes.  Currently
this is used to carry the IP address of the client.  Some doors send
this information some of the time.

This patch updates doors to always send this information, with the
exception of the NFS door.  This door uses gPlazma only when logging
in GSS-based connections, but currently the interface does not
provide the client IP address.  This will be fixed with future
patches.

Target: master
Require-notes: no
Require-book: no
Patch: https://rb.dcache.org/r/8825
Acked-by: Gerd Behrmann