chimera: Read AL and RP on directory listing
Motivation:

In 2.14 we embedded the access latency and retention policy in t_inodes and
thus in the Stat object. We fail to populate the access latency and retention
policy in the Stat object during directory listing.

This oversight is a direct consequence of code duplication in Chimera.

Modification:

Refactor the t_inodes to Stat mapping such that it can be reused between the
code doing listing and the implementation of the stat call.

Result:

Correct access latency and retention policy in directory listings (propably mostly
something WebDAV and SRM care about).

Target: trunk
Request: 2.14
Require-notes: yes
Require-book: no
Acked-by: Paul Millar <paul.millar@desy.de>
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: https://rb.dcache.org/r/8844/