chimera: When updating stat cache, take the file type into account
Motivation:

When updating attributes of a file, the in memory stat cache is updated with
the new values. The stat structure however overloads the mode field to be the
union of the Chimera imode and itype fields and the update method fails to take
this into account. Thus after updating the mode of eg a directory, a call to
FsInode#isDirectory returns false afterwards.

Modification:

Preserve the type. It is assumed that we would never change the type of a
Chimera inode.

Result:

It is not clear whether any of the existing code was affected by this issue.

Target: trunk
Request: 2.14
Require-notes: yes
Require-book: no
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: https://rb.dcache.org/r/8847/