chimera: Prevent filling of stat cache of root inode
Motivation:

The Chimera FsInode object can cache stat content of a t_inodes row. Such information
may be invalidated by concurrent operations, but since FsInode instances are usually
short-lived often local to a single transaction, this isn't normally a problem.

There is however one FsInode instance that has a longer life time: The one representing
the file system root allocated in the JdbcFs class. Currently this instance may
cache the stat information indefinitely and updates on the root inode (even if rare)
would not be visible until the next restart.

Modification:

Copy the root FsInode in any operation that may return it to the caller of JdbcFs or
which may access the stat information and thus trigger the population of the cache.

Introduce a subclass of FsInode specifically for the root inode: Some properties we
know even without accessing the t_inodes table, such as that it is a directory and
that it exists.

Result:

Correct information about the root directory, but unfortunately at the price of
an increase in stat calls.

Target: trunk
Require-notes: yes
Require-book: no
Request: 2.14
Request: 2.13
Request: 2.12
Request: 2.11
Request: 2.10
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: https://rb.dcache.org/r/8849/