chimera: Add inumber identity field to t_inodes
Motivation:

PNFSIDs serve as permanent identifiers for files and directories in dCache, and
with the exception of legacy IDs imported from PNFS are globally unique.

The price is that the ID is a 36 character string - a string that as far as
postgresql is concerned doesn't have a fixed length, i.e. it's actually 37
bytes long, which needs to be aligned on 4 byte boundaries, so it is likely
more like 40 bytes.

Since we use PNFS ID as a primary key, this overhead is also payed by indexes,
foreign keys, and indexes on foreign keys. The number of elements in a b-tree
node is reduced, growing the height of the tree, and variable string
comparisons are fairly expensive, at least compared to CPU native integers.

Modification:

This patch introduces an 8 byte identity column called inumber in t_inodes.
The field is filled automatically by the RDBMS by a sequence.

All foreign keys are updated to use inumber rather than ipnfsid. Upon upgrade
all Chimera tables other than t_inodes are effectively rewritten.

The FsInode object is updated to tracket inumber rather than ipnfsid. Whereas
Chimera previously used a hash of PNFS ID as an inode number, inumber now
serves this purpose. The opaque encoding of an FsInode used by NFS now encodes
inumber rather than ipnfsid.

The Stat object is extended with a field tracking the PNFS ID. I.e. the
persistent id is now an attribute of the file like any other attribute.

A cache for mapping PNFS ID to inumber and vice versa is introduced.

The inumber is stable although not globally unique. The only use case in which
the inumber would change is when two dCache instances are merged. For this
reason the inumber should not be used for long term reference outside of
Chimera, but it might serve quite fine for short term identifiers e.g. for NFS
(if two dCache instances are merged, it is a safe bet that both instance will
be shut down and all NFS mounts will be unmounted).

Stored procedures are updated to operate on inumber, but since inode2path,
inode2paths, and path2inode are commonly used in third party scripts, these
are kept with the existing semantics.

Field order in several tables is changed to put variable length fields last
and fixed length non-nullable fields first; this results in better packing
and pre-computable offsets in postgresql.

For the checksum table, the checksum type is now part of the primary key,
allowing a file to have multiple checksums of different types.

Result:

Upon upgrade the schema is modified to add an inumber identity column to
t_inodes. In contrast to the PNFS ID, the inumber is only unique within this
instance and not globally unique. It is generated sequentially from a database
sequence. Foreign keys of other Chimera tables are updates to use inumber
instead.

If two dCache instances are to be merged, the inumber values of one need to be
updated to aviod collisions. Specifically, one determines the maximum from
instance A (select max(inumber) from a.t_inodes), the minimum from instance B
(select min(inumber) from b.t_inodes) and then updates B to avoid collisions
with A (update b.t_inodes set inumber = inumber + max_a - min_b + 1). After this
the two instances are mergable. Since the PNFS ID didn't change, no files on
pools or tapes need to be updated.

Target: master
Require-notes: yes
Require-book: yes
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: https://rb.dcache.org/r/8986/