Add ZooKeeper support
Motivation:

ZooKeeper is a distributed coordination service with strong ordering
guarantees. It is commonly used for service discovery, leader election,
configuration management in distributed systems, etc.

Apache Curator is a utility library that acts as a facade to ZooKeeper.

Modification:

Adds a ZooKeeper client for each domain. The client is accessible through the
Apache Curator framework. Each CellNucleus wraps the Curator framework instance
to mainain the CDC and to use the cell's message executor as a default executor
for Curator callbacks and listeners.

CellShell is extended with the `zk ls` and `zk get` commands to list and query
ZooKeeper nodes.

A zookeeper service is added to embed a single-node ZooKeeper inside dCache.
This is mostly for purposes of testing, evaluation and migration. Any
production instance SHOULD use an independent ZooKeeper instance (either
single-node or multi-node).

The client defaults to connect to the broker host, thus if the zookeeper
service is embedded into dCacheDomain, no further configuration changes are
needed.

Result:

Nothing useful, but this will serve as a foundation for using ZooKeeper in
dCache.

Since the ZooKeeper instance is shared for all services in a domain, care
should be taken that a service removes any ephemeral nodes it created. This is
because a dCache service may terminate without the domain shutting down.  Thus
one cannot rely on ZooKeeper itself removing such nodes (in case of a crash or
session expiration, the nodes will of course still be removed).

Target: trunk
Require-notes: yes
Require-book: no
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: https://rb.dcache.org/r/8854/