dcache-webdav: Moved handling of authentication from Milton Security Filter to it separate handler
Motivation:

Currently, the authentication of incoming requests to Webdav module occurs in the Security Filter of the Milton Handler. In order to make the authentication also available to other handlers, it was important that it happens outside of the filter chain.

Ticket:
Acked-by: Paul Millar <paul.millar@desy.de>
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Target: trunk
Require-book: no
Require-notes: no
Patch: https://rb.dcache.org/r/8870/