pool: Add migration option to filter by absense of sticky flags
Motivation:

A regression caused sticky bits to be lost. To repair this, one could
use a migration job that specifically copies cached and not sticky disk
files to other pools and in the process sets the sticky bit. To allow
this one needs to be able to filter by files which are cached and not
sticky and this is currently not possible.

Modification:

Extends the -sticky option with the ability to negate the filter.

Result:

One can filter for all files not having the system sticky flag by
using -sticky=-system.

To allow sites to repair the recent bug, backporting the change is
requested.

Target: trunk
Request: 2.14
Request: 2.13
Request: 2.12
Require-notes: yes
Require-book: no
Acked-by: Albert Rossi <arossi@fnal.gov>
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: https://rb.dcache.org/r/8872/