Add caching layer for TLS validation
Motivation:

The CANL validator is single threaded, causing the cert chain validation after
a TLS handshake to become a bottleneck.

Modification:

Adds a caching layer on top of CANL. The entire chain is hashed and the result
of the validation is cache using the hash as the key. The lifetime of the
certificates of the chain is always checked even when using a cached result.

The default cache time is 5 minutes and is currently not configurable. This
will likely change in a future version.

The VOMS library provides a similar caching layer, but it only hashes on the
last certificate of the chain. This provides an opportunity for an attacker to
replace the rest of the chain, eg. to inject a different EEC.

Result:

Hopefully better TLS performance.

Target: trunk
Request: 2.14
Require-notes: yes
Require-book: no
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: https://rb.dcache.org/r/8876/
(cherry picked from commit 482028c4f98fb4944bfa5ea86c092813ea2509a6)