pool: Do not output expired sticky flags
Motivation:

The way we clear a sticky flag is to set it to an expiration date in the past -
it will expire right away. A consequence of this is that if the sticky flags are
output right after being set/cleared, a sticky flag with expiration at the Unix
epoch is included.

Modification:

Filter expired sticky flags in the output of 'rep ls sticky' and 'rep set sticky'.

Result:

Less confusing output from these commands.

Target: trunk
Require-notes: no
Require-book: no
Request: 2.14
Request: 2.13
Request: 2.12
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: https://rb.dcache.org/r/8883/