info: fix test to be less critical on timing
The unit-tests for info check whether components behave correctly
according to their expiry date.  One such test checks whether the expiry
date is as expected.  Metric expiry dates are rounded up to the nearest
half-second, to increase the likelihood of expunging multiple metrics
that were created at about the same time.

Unfortunately, the test has a simple tolerance set to the rounding
factor (500 ms).  If the object is created a little before a half-second
boundary and the test is made after this boundary then the result can
easily exceed this threshold, so triggering a false failure of the
unit-test.

This patch updates the test to take this into account.  It is still a
work-around, as the correct solution is to decouple info from the time
source, to allow for better control over its behaviour.  However, this
should be simple enough to back-port safely.

Target: master
Request: 2.14
Request: 2.13
Request: 2.12
Request: 2.11
Request: 2.10
Requires-notes: no
Requires-book: no
Patch: https://rb.dcache.org/r/8886/
Acked-by: Albert Rossi