http-3rd-party: ensure IOException logged with toString
The pool's support for 3rd party transfers is written to accept
the possibility of failure due to IOException.  Unfortunately,
it is not guaranteed that such exceptions will contain any
message, which means that using the getMessage method may
be insufficient to learn what caused the problem.

This patch updates the HTTP 3rd-party handling code to log
both the exception type and the message (if any) for any
IOException.

Target: master
Require-notes: yes
Require-book: no
Acked-by: Gerd Behrmann
Request: 2.14
Request: 2.13
Request: 2.12
Request: 2.11
Request: 2.10

Conflicts:
	modules/dcache/src/main/java/org/dcache/pool/movers/RemoteHttpDataTransferProtocol_1.java