srm-client: Fix GSI delegation for old servers
Motivation:

During GSI delegation JGlobus expects the client to send the entire certificate
in a single SSL frame. A workaround for the TLS BEAST attack makes the JVM
split the first payload into two chunks, the first being 1 byte long. This
conflicts with the expectations of JGlobus and thus causes our 2.14 srm client
to fail with dCache older than 2.13. This was not a problem while the client
used JGlobus and thus didn't use protection against BEAST.

Modification:

Disable the protection against the BEAST attack in the SRM client when
delegating credentials to the server. This fixes the delegation with GSI at the
expense of vulenrability to BEAST if the server generates HTTPS TURLs.

Result:

GSI delegation works with JGlobus.

Target: trunk
Request: 2.14
Require-notes: yes
Require-book: no
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: https://rb.dcache.org/r/8896/