pool: Fix health check of file store
Motivation:

Both the file and meta data stores support health checks, however they were
only called for the meta data store. This is because it is the ConsistentStore
that binds the two stores together, but the health checker called into the low
level meta store directly.

Modification:

"Lift" the ConsistentStore in the spring file such that it is injected both
into the repository and the health checker. Also moved the call to close into
the repository as the repository is wrapping the store with a cache and the
cache should be closed too.

Result:

Periodic health checks for the data directory.

Target: trunk
Request: 2.14
Request: 2.13
Require-notes: yes
Require-book: no
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: https://rb.dcache.org/r/8911/