pool: Do not fail read if xrootd client doesn't close file
Motivation:

Xrootd has a flag for POSC semantics:

    The kXR_posc option requests safe file persistence which persists the file
    only when it has been explicitly closed.

In can be concluded that for files opened with this flag, disconnecting without
closing the file is not really an error.

Modification:

Do not generate an error if files being read are not explicitly closed.

We still generate an error for writes without a close. dCache persists such
files so semantically it is in line with the xrootd specification. Generating
an error in these cases may help recognize failing clients.

Result:

No more

    666:"Client disconnected without closing file."

errors on xrootd reads.

Target: trunk
Request: 2.14
Request: 2.13
Require-notes: yes
Require-book: no
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: https://rb.dcache.org/r/8920/