pool: Redirect xrootd client to door on failure to open file
Motivation:

Xrootd clients have to provide a UUID obtained from the door when opening a
file on a pool. If no UUID is included or if it is no longer valid we currently
log an error message and return an error to the client.

Modification:

Attempt to redirect the client to a door - this is only possible if the
client already opened another file as we otherwise do not know the door
to redirect the client to.

Reduce the log level of the message logged to info as this is a relatively
common occurence and there is no reason to spam our log files.

Avoid logging the entire request as it may contain a confidential authorization
token.

Result:

Reduced log spam on pools in case xrootd clients fail to present a valid UUID.

Target: trunk
Request: 2.14
Request: 2.13
Require-notes: yes
Require-book: no
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: https://rb.dcache.org/r/8921/