gplamza: update LoginResultPrinter to show more AC information
Proxy certificates include an embedded attribute certificate when they
are a voms proxy.  This attribute certificate contains group-membership
information and is signed by the VOMS server's credential.

An .lsc file associates a trusted VOMS server with some already trusted
certificate authority.  This requires that the CA that signed the VOMS
server's certificate matches the DN in the .lsc file, otherwise the
VOMS server is not trusted and no FQANs are extracted.  Therefore, when
diagnosing a problem where the voms plugin fails to extract any FQANs
it is sometimes useful to know which CA signed the VOMS server's
certificate.

This patch updates the LoginResultPrinter to show the attribute
certificate extensions.  One VOMS-specific extension includes the
certificate chain of the VOMS server.  With this patch, the various
issuers of this chain are shown.

In most cases the certificate of the VOMS server is signed by some root
CA; i.e., the chain has length one.  In this case a slightly more
optimal format is used to show the chain.

Additionally, the output of the DN of the VOMS certificate is updated
to RFC 2253 format.

Target: master
Requires-notes: yes
Requires-book: yes
Patch: https://rb.dcache.org/r/8926/
Acked-by: Gerd Behrmann