webdav: minor refactor
Remove the reverse-mapping in favour of using the Milton-supplied path
when establishing the original request path.  This path also changes
dCache behaviour should Milton supply a malformed URL: previously this
is logged, with this patch it is treated as a bug.

Target: master
Require-notes: no
Require-book: no
Patch: https://rb.dcache.org/r/8943
Acked-by: Tigran Mkrtchyan