pool: Refactor pool initialization to reduce exclusive locking
Motivation:

During pool startup we have to ensure that the initial "load event" for an
entry is generated before any change events made concurrently with pool
initialization. This is complicated by the fact that the entry may be triggered
by a client read rather than the thread initializing the repository.

The present solution was to suppress state change notifications for entries
until the repository was initialized; thus the initialization procedure would
generate all the load events and only then would repository changes generate
change events. For this to work an exclusive lock had to be held for the final
loop in which the load events were generate. For large pools, this could still
take considerable time (minutes) in which all other access to the pool would be
blocked.

This patch tries to address this problem.

Modification:

The strategy is to generate the load event as soon as an entry is loaded, thus
removing the need to suppress subsequent change events.

This is achieved by modifying the meta data cache. The cache already
synchronizes access to each entry and maintaines a monitor for each entry. This
monitor object is extended to act as a decorator for the entry, providing a
perfect place to generate change events.

In contrast to the state change events of the main repository, the cache calls
its listener synchronously and with the meta data entry locked. This allows the
sole listener, CacheRepositoryV5, to synchronize its state and the state of its
collaborator with the meta data entries. In particular it can update the
Account and schedule sticky flag expiration tasks. Other listeners are notified
as before from a separate thread and without locking the meta data entry.

Since the cache triggers the state change event it doesn't matter if the
initial load was triggered by the initialization thread or a client read.

One consequence of this change is that the account needs to be updated on the
fly while loading entries. The Account class is extended to allow this.

Result:

Less lock contention on pool initialization. In particular the period of
exclusive access at the end of pool intiialzation is gone.

Target: trunk
Require-notes: yes
Require-book: no
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: https://rb.dcache.org/r/8948/