webdav: log client-chain based on X-Forwarded-For header
Motivation:

If an HTTP/WebDAV request is made via some proxy then dCache provides no
information about the client IP address: only the proxy IP address is
logged.  This is undesirable in circumstances where an admin wishes to
establish the identity of the client initiating the request.

Modification:

Support is added for the X-Forwarded-For HTTP request header.  This is
the de facto standard approach to log the client chain.  However, it
provides no guarantee of veracity; therefore any address value is only
as trustworthy as the subsequent proxy services (if any).

Result:

The text billing output is updated to log the client-chain.  For simple
(non-proxied) requests there is no change, but proxied requests show a
comma-separated list of IP addresses.  The old single-valued attribute
is still available for backwards compatibility.

Target: master
Requires-notes: yes
Requires-boook: yes
Patch: https://rb.dcache.org/r/8955/
Acked-by: Tigran Mkrtchyan